Структура гео-классов

+++ src/main/java/ru/griga/tickets/model/Itinerary.java
@@ -2,6 +2,9 @@ package ru.griga.tickets.model;
 
 import lombok.Data;
 import org.neo4j.ogm.annotation.*;
+import ru.griga.tickets.model.place.base.Place;
+
+import java.math.BigDecimal;
 
 @Data
 @RelationshipEntity(type = "CAN_GO")
@@ -17,6 +20,6 @@ public class Itinerary {
     @EndNode
     private final Place destination;
 
-    private final float cost;
+    private final BigDecimal cost;
 
 }
